mergeqqq

+++ .buildkite/pipeline.sh
@@ -2,11 +2,27 @@
 
 echo "BUILDKITE_BRANCH: $BUILDKITE_BRANCH"
 
-cat <<EOF | tee /dev/tty | buildkite-agent pipeline upload
-steps:
-  - name: "main"
-    command: "./check.sh"
-    timeout_in_minutes: 20
-    agents:
-      queue: "default"
+if [[ $BUILDKITE_BRANCH == gh-readonly-queue* ]]; then
+  cat <<EOF | tee /dev/tty | buildkite-agent pipeline upload
+  steps:
+    - name: "main"
+      command: "./check.sh"
+      timeout_in_minutes: 20
+      agents:
+        queue: "default"
+    - name: "main"
+      command: "./check.sh"
+      timeout_in_minutes: 20
+      agents:
+        queue: "default"
+EOF 
+else
+  cat <<EOF | tee /dev/tty | buildkite-agent pipeline upload
+  steps:
+    - name: "main"
+      command: "./check.sh"
+      timeout_in_minutes: 20
+      agents:
+        queue: "default"
 EOF
+fi
